# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "材料实验助手 API"
  version: "1.0.0"
  description: |-
    当前后端实现的接口清单；维护依据为 backend 中的 Controller、Service、响应模型和安全配置。
    普通成功响应使用 data 和 request_id；分页另有 meta。空 data 按当前 NON_NULL 配置省略。
    204 和文件正文不使用成功信封。错误返回 application/json，包含业务 Problem Details 兼容对象、精简安全错误或内部错误；不能假定所有错误都有 detail、instance 或 request_id。

    前端使用同源 Session（默认 Cookie 为 JSESSIONID）。先获取 /auth/csrf 并保留 csrftoken，写请求携带匹配的 X-CSRFToken；登录也受 CSRF 校验。当前 SecurityConfig 另启用了 HTTP Basic。
    项目和实验读取按当前组织限定数据范围，普通用户另受项目、负责人或实验参与关系约束。创建实验或变更实验关联项目时先校验项目属于当前组织；超级管理员仅跳过项目管理成员检查。平台管理员仅能开通与查看组织目录，不获得其他组织的业务数据读取权。项目和实验受版本保护的操作使用 ETag / If-Match；当前没有 Idempotency-Key 去重实现。

    时间字段沿用现有 API 的 ISO 8601 表示，带时间的响应为含偏移量的日期时间。规范要求客户端展示及人工记录统一到 Asia/Shanghai，但当前页面部分格式化使用浏览器默认时区，尚未统一。项目和实验计划时间省略偏移量时服务端按 Asia/Shanghai 解释；调用方仍应优先提交含偏移量的日期时间。

    能力边界：ready 检查应用启动就绪及数据库连接；项目文档和实验附件正文保存在 PostgreSQL，文档 preview 返回原件。RustFS 存储和 LibreOffice 服务端转换尚未接入。
    实验参与人可随创建或 PATCH 写入；项目成员仅可读取，创建时加入创建者。开发初始化不能替代生产组织、角色和管理员引导；本契约不表示已完成生产部署验收。
servers:
  -
    url: "/api/v1"
    description: "同源 API"
...
security:
  -
    cookieAuth: []
  -
    basicAuth: []
paths:
...
  "/auth/csrf":
    get:
      operationId: "getCsrfToken"
      tags: ["Auth"]
      summary: "获取 CSRF 令牌"
      security: []
      responses:
        "200":
          description: "返回令牌并设置或复用 csrftoken Cookie"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/Envelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "object"
                        required: ["csrf_token"]
                        properties:
                          csrf_token:
                            type: "string"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/login":
    post:
      operationId: "login"
      tags: ["Auth"]
      summary: "登录并建立服务器会话"
      description: "登录也受 CSRF 校验；username 和 password 必须非空白字符串。"
      security:
        -
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: "object"
              required: ["username","password"]
              properties:
                username:
                  type: "string"
                  minLength: 1
                password:
                  type: "string"
                  minLength: 1
                  format: "password"
                  writeOnly: true
      responses:
        "200":
          "$ref": "#/components/responses/Session"
        "401":
          "$ref": "#/components/responses/ApiError"
        "403":
          "$ref": "#/components/responses/ApiError"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/register":
    post:
      operationId: "registerByInvitation"
      tags: ["Auth"]
      summary: "使用一次性邀请码注册普通用户"
      description: "无需登录但必须携带 CSRF。邀请码决定组织和角色，明文邀请码不被持久化；注册不自动登录，不能创建超级管理员。"
      security:
        -
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/InvitationRegistration"
      responses:
        "201":
          "$ref": "#/components/responses/EmptySuccess"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/logout":
    post:
      operationId: "logout"
      tags: ["Auth"]
      summary: "失效当前服务器会话"
      description: "按同源 Session 流程调用，要求已有服务器会话。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
      responses:
        "204":
          description: "注销成功，无响应正文"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/password":
    post:
      operationId: "changeOwnPassword"
      tags: ["Auth"]
      summary: "修改当前登录用户密码"
      description: "校验当前密码与统一密码策略（至少 8 个字符，包含字母和数字，不含空白或控制字符且不得与用户名相同）；成功后立即失效当前 Servlet Session，调用方需重新登录。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/OwnPasswordChange"
      responses:
        "204":
          description: "密码已修改，当前会话已失效"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/session":
    get:
      operationId: "getSession"
      tags: ["Auth"]
      summary: "获取当前认证用户"
      responses:
        "200":
          "$ref": "#/components/responses/Session"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/users/options":
    get:
      operationId: "listUserOptions"
      tags: ["Auth"]
      summary: "查询当前组织及下级组织的有效用户选项"
      description: "要求 project.create 权限；递归包含当前组织及下级组织的 active 用户，仅返回 id、username 和 display_name。注意项目和实验负责人写入仍要求用户属于当前组织，因此下级组织选项不一定可用于保存。"
      responses:
        "200":
          description: "有效用户选项"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/Envelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/UserOption"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/roles/options":
    get:
      operationId: "listRoleOptions"
      tags: ["Auth"]
      summary: "查询当前组织可分配角色"
      description: "仅超级管理员可访问；角色代码以本接口返回值为准。"
      responses:
        "200":
          description: "角色选项"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/Envelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/RoleOption"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/invitations":
    get:
      operationId: "listRegistrationInvitations"
      tags: ["Auth"]
      summary: "查询当前组织的邀请码元数据"
      description: "仅超级管理员。邀请码明文仅在创建响应中返回一次，列表不包含明文或哈希。"
      responses:
        "200":
          description: "最近 100 条邀请码元数据"
          content:
            "application/json":
              schema:
                allOf:
                  - "$ref": "#/components/schemas/Envelope"
                  - type: "object"
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/RegistrationInvitation"
        default:
          "$ref": "#/components/responses/ApiError"
    post:
      operationId: "createRegistrationInvitation"
      tags: ["Auth"]
      summary: "签发绑定角色的一次性邀请码"
      description: "仅超级管理员。角色不能是 super_admin；邀请码有效期为 1 至 720 小时，明文只随本次响应返回。"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/RegistrationInvitationCreate"
      responses:
        "201":
          description: "邀请码签发成功"
          content:
            "application/json":
              schema:
                allOf:
                  - "$ref": "#/components/schemas/Envelope"
                  - type: "object"
                    properties:
                      data:
                        "$ref": "#/components/schemas/RegistrationInvitationIssue"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/invitations/{invitationId}/revoke":
    post:
      operationId: "revokeRegistrationInvitation"
      tags: ["Auth"]
      summary: "撤销尚未使用的邀请码"
      description: "仅超级管理员；邀请码已使用、已撤销或不存在时返回 404。"
      parameters:
        -
          name: "invitationId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "204":
          description: "邀请码已撤销"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/users":
    get:
      operationId: "listUsers"
      tags: ["Auth"]
      summary: "分页查询当前组织用户"
      description: "仅超级管理员可访问。"
      parameters:
        -
          "$ref": "#/components/parameters/Page"
        -
          "$ref": "#/components/parameters/PageSize"
        -
          "$ref": "#/components/parameters/Search"
        -
          name: "status"
          in: "query"
          schema:
            type: "string"
          description: "用户状态筛选，例如 active、locked、disabled"
        -
          name: "role_code"
          in: "query"
          schema:
            type: "string"
          description: "角色选项接口返回的角色代码"
      responses:
        "200":
          description: "用户分页列表"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/PageEnvelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/ManagedUser"
        default:
          "$ref": "#/components/responses/ApiError"
    post:
      operationId: "createUser"
      tags: ["Auth"]
      summary: "创建当前组织用户并绑定角色"
      description: "仅超级管理员。当前返回 HTTP 200 和 request_id，不返回新建用户实体。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/UserCreate"
      responses:
        "200":
          "$ref": "#/components/responses/EmptySuccess"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/users/{userId}":
    parameters:
      -
        "$ref": "#/components/parameters/UserId"
    patch:
      operationId: "updateUser"
      tags: ["Auth"]
      summary: "更新当前组织普通用户"
      description: |-
        仅超级管理员，不能修改超级管理员账号。虽使用 PATCH，username、display_name、status、role_codes 均必须完整提交。
        email 省略时写入空串，角色关联按本次 role_codes 整体替换。返回 request_id，不返回用户实体。
        当前没有按用户 ID 读取详情的 GET 接口。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/UserWrite"
      responses:
        "200":
          "$ref": "#/components/responses/EmptySuccess"
        default:
          "$ref": "#/components/responses/ApiError"
  "/auth/users/{userId}/reset-password":
    parameters:
      -
        "$ref": "#/components/parameters/UserId"
    post:
      operationId: "resetUserPassword"
      tags: ["Auth"]
      summary: "重置当前组织普通用户密码"
      description: "仅超级管理员；超级管理员账号受保护；密码至少 8 个字符，包含字母和数字，不含空白或控制字符且不得与用户名相同。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              type: "object"
              required: ["password"]
              properties:
                password:
                  type: "string"
                  minLength: 8
                  format: "password"
                  writeOnly: true
      responses:
        "204":
          description: "密码已更新，无响应正文"
        default:
          "$ref": "#/components/responses/ApiError"
...
components:
  securitySchemes:
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "JSESSIONID"
      description: "Servlet Session 默认 Cookie 名；服务器配置可覆盖"
    basicAuth:
      type: "http"
      scheme: "basic"
      description: "当前 SecurityConfig 启用的 HTTP Basic；写请求仍受 CSRF 约束"
    csrfCookie:
      type: "apiKey"
      in: "cookie"
      name: "csrftoken"
      description: "先调用 GET /auth/csrf 获取并保留"
    csrfHeader:
      type: "apiKey"
      in: "header"
      name: "X-CSRFToken"
      description: "与当前 csrftoken Cookie 匹配的令牌"
  parameters:
    Page:
      name: "page"
      in: "query"
      schema:
        type: "integer"
        default: 1
      description: "页码从 1 开始，低于 1 按 1 处理"
    PageSize:
      name: "page_size"
      in: "query"
      schema:
        type: "integer"
        default: 20
      description: "每页数量，服务端归一化到 1 至 100"
    Search:
      name: "search"
      in: "query"
      schema:
        type: "string"
      description: "关键字模糊筛选，具体搜索列以接口查询实现为准"
    UserId:
      name: "userId"
      in: "path"
      required: true
      schema:
        type: "string"
        format: "uuid"
      description: "当前组织用户主键"
...
  responses:
    ApiError:
      description: "业务或参数错误为 Problem Details 兼容对象；安全过滤器返回 status/code/detail；未分类异常返回内部错误。通常 401 为未认证、403 为权限或 CSRF 失败、404 为资源不可见或不存在。当前媒体类型为 application/json，不能假定所有非法输入均为 400。"
      content:
        "application/json":
          schema:
            anyOf:
              -
                "$ref": "#/components/schemas/ProblemDetail"
              -
                "$ref": "#/components/schemas/SecurityError"
              -
                "$ref": "#/components/schemas/InternalError"
    EmptySuccess:
      description: "操作成功；当前 NON_NULL 配置省略空 data，仅返回 request_id"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Envelope"
    Session:
      description: "当前认证用户及权限"
      content:
        "application/json":
          schema:
            allOf:
              -
                "$ref": "#/components/schemas/Envelope"
              -
                type: "object"
                required: ["data"]
                properties:
                  data:
                    "$ref": "#/components/schemas/SessionUser"
...
  schemas:
    Envelope:
      type: "object"
      required: ["request_id"]
      properties:
        request_id:
          type: "string"
          description: "当前响应生成的标识"
          format: "uuid"
      description: "非空业务数据通过 data 返回；空 data 按当前 NON_NULL 配置省略。"
    PageEnvelope:
      allOf:
        -
          "$ref": "#/components/schemas/Envelope"
        -
          type: "object"
          required: ["meta"]
          properties:
            meta:
              "$ref": "#/components/schemas/PageMeta"
    PageMeta:
      type: "object"
      required: ["page","page_size","total","total_pages"]
      properties:
        page:
          type: "integer"
          minimum: 1
        page_size:
          type: "integer"
          minimum: 1
          maximum: 100
        total:
          type: "integer"
          format: "int64"
          minimum: 0
        total_pages:
          type: "integer"
          format: "int64"
          minimum: 1
          description: "即使 total 为 0 也返回 1"
    ProblemDetail:
      type: "object"
      required: ["type","title","status","code","detail","instance","request_id"]
      properties:
        type:
          type: "string"
          const: "about:blank"
        title:
          type: "string"
          description: "HTTP 状态英文名称"
        status:
          type: "integer"
        code:
          type: "string"
          description: "业务错误代码，例如 version_conflict"
        detail:
          type: "string"
          description: "业务错误或校验说明"
        instance:
          type: "string"
          description: "请求 URI"
        request_id:
          type: "string"
          format: "uuid"
    SecurityError:
      type: "object"
      required: ["status","code","detail"]
      properties:
        status:
          type: "integer"
          enum: [401,403]
        code:
          type: "string"
          enum: ["authentication_required","permission_denied"]
        detail:
          type: "string"
          const: "认证或权限校验失败"
      description: "安全过滤器直接返回，不包含 request_id、type 或 instance。"
    InternalError:
      type: "object"
      required: ["type","title","status","code","request_id"]
      properties:
        type:
          type: "string"
          const: "about:blank"
        title:
          type: "string"
          const: "服务器内部错误"
        status:
          type: "integer"
          const: 500
        code:
          type: "string"
          const: "internal_error"
        request_id:
          type: "string"
          format: "uuid"
      description: "未分类异常的兜底响应，不包含 detail 或 instance。"
...
    SessionUser:
      type: "object"
      required: ["id","username","display_name","organization_id","is_super_admin","is_platform_admin","permissions","role_codes","role_names"]
      properties:
        id:
          type: "string"
          format: "uuid"
        username:
          type: "string"
        display_name:
          type: "string"
        organization_id:
          type: "string"
          format: "uuid"
        is_super_admin:
          type: "boolean"
        is_platform_admin:
          type: "boolean"
          description: "仅控制组织目录与开通权限；不扩大当前会话的业务数据范围"
        permissions:
          type: "array"
          items:
            type: "string"
          description: "超级管理员返回单项 *，普通用户返回具体权限代码"
        role_codes:
          type: "array"
          items:
            type: "string"
        role_names:
          type: "array"
          items:
            type: "string"
    UserOption:
      type: "object"
      required: ["id","username","display_name"]
      properties:
        id:
          type: "string"
          format: "uuid"
        username:
          type: "string"
        display_name:
          type: "string"
    RoleOption:
      type: "object"
      required: ["role_code","name"]
      properties:
        role_code:
          type: "string"
        name:
          type: "string"
        description:
          type: "string"
    InvitationRegistration:
      type: "object"
      required: ["invitation_code", "username", "display_name", "password"]
      properties:
        invitation_code:
          type: "string"
          writeOnly: true
        username:
          type: "string"
          minLength: 1
        display_name:
          type: "string"
          minLength: 1
        email:
          type: "string"
        password:
          type: "string"
          minLength: 8
          description: "至少 8 个字符，最多 72 个 UTF-8 字节；包含字母和数字，不含空白或控制字符且不得与用户名相同。"
          format: "password"
          writeOnly: true
    OwnPasswordChange:
      type: "object"
      required: ["current_password", "new_password"]
      properties:
        current_password:
          type: "string"
          format: "password"
          writeOnly: true
        new_password:
          type: "string"
          minLength: 8
          description: "至少 8 个字符，最多 72 个 UTF-8 字节；包含字母和数字，不含空白或控制字符且不得与用户名相同。"
          format: "password"
          writeOnly: true
    RegistrationInvitationCreate:
      type: "object"
      required: ["role_code"]
      properties:
        role_code:
          type: "string"
          description: "当前组织有效角色代码，不能为 super_admin"
        valid_for_hours:
          type: "integer"
          minimum: 1
          maximum: 720
          default: 168
    RegistrationInvitationIssue:
      type: "object"
      required: ["invitation_code", "role_code", "expires_at"]
      properties:
        invitation_code:
          type: "string"
          description: "仅本次签发响应包含；不得记录或再次查询"
        role_code:
          type: "string"
        expires_at:
          type: "string"
          format: "date-time"
    RegistrationInvitation:
      type: "object"
      required: ["id", "role_code", "role_name", "status", "expires_at", "created_at"]
      properties:
        id:
          type: "string"
          format: "uuid"
        role_code:
          type: "string"
        role_name:
          type: "string"
        status:
          type: "string"
          enum: ["active", "used", "revoked", "expired"]
          description: "expired 为列表读取时依据过期时刻计算的展示状态，不写入数据库"
        expires_at:
          type: "string"
          format: "date-time"
        used_at:
          type: "string"
          format: "date-time"
        created_at:
          type: "string"
          format: "date-time"
    UserStatus:
      type: "string"
      enum: ["active","locked","disabled"]
    UserWrite:
      type: "object"
      required: ["username","display_name","status","role_codes"]
      properties:
        username:
          type: "string"
          minLength: 1
        display_name:
          type: "string"
          minLength: 1
        email:
          type: "string"
          description: "省略时写入空串；当前未执行邮箱格式校验"
        status:
          "$ref": "#/components/schemas/UserStatus"
        role_codes:
          type: "array"
          minItems: 1
          items:
            type: "string"
          description: "当前组织有效角色代码，按整组替换"
    UserCreate:
      allOf:
        -
          "$ref": "#/components/schemas/UserWrite"
        -
          type: "object"
          required: ["password"]
          properties:
            password:
              type: "string"
              minLength: 8
              description: "至少 8 个字符，最多 72 个 UTF-8 字节；包含字母和数字，不含空白或控制字符且不得与用户名相同。"
              format: "password"
              writeOnly: true
    ManagedUser:
      type: "object"
      required: ["id","username","display_name","status","active","super_admin","role_codes","role_names","created_at","updated_at"]
      properties:
        id:
          type: "string"
          format: "uuid"
        username:
          type: "string"
        display_name:
          type: "string"
        email:
          type: "string"
        status:
          "$ref": "#/components/schemas/UserStatus"
        active:
          type: "boolean"
        super_admin:
          type: "boolean"
          description: "用户列表字段名；会话中对应字段为 is_super_admin"
        role_codes:
          type: "array"
          items:
            type: "string"
        role_names:
          type: "array"
          items:
            type: "string"
        last_login:
          type: "string"
          description: "无值时省略；当前登录服务未更新此字段"
          format: "date-time"
        created_at:
          type: "string"
          format: "date-time"
        updated_at:
          type: "string"
          format: "date-time"
...